requirements.txt
bcrypt==3.2.0
blinker==1.9.0
certifi==2020.6.20
chardet==4.0.0
cryptography==3.4.8
Flask==2.3.2          # Cambié la versión a una estable de Flask
Flask-SocketIO==5.3.4
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==8.1.0
python-dotenv==1.0.1
requests==2.25.1
six==1.16.0
Werkzeug==3.1.3
eventlet==0.33.0
greenlet==2.0.2
gunicorn==20.1.0
